Move exp, bigint to test.
