+++ src/main/kotlin/org/cryptobiotic/gmp/EgkGmp.kt
@@ -1,9 +1,5 @@
 package org.cryptobiotic.gmp
 
-import electionguard.core.*
-import org.cryptobiotic.gmp.EgkGmpIF
-import org.cryptobiotic.mixnet.VectorCiphertext
-import org.cryptobiotic.mixnet.VectorQ
 import java.lang.foreign.*
 import java.lang.foreign.ValueLayout.*
