+++ src/main/kotlin/org/cryptobiotic/mixnet/ShuffleProver.kt
@@ -271,92 +271,6 @@ data class ProofCommittment (
     val Fp: VectorCiphertext, // width
 )
 
-/////////////////////////////////////////////////////////////////////////////////////////
-
-// componentwise product of column vectors to a power
-// return VectorCiphertext(width)
-// CE (2 exp) N
-// TODO is this really what vmn does?
-fun prodColumnPow(rows: List<VectorCiphertext>, exps: VectorQ, nthreads: Int = 10): VectorCiphertext {
-    return if (nthreads == 0) {
-        prodColumnPowSingleThread(rows, exps)                                                // CE 2 * N exp
-    } else {
-        PprodColumnPow(rows, exps, nthreads).calc()
-    }
-}
-
-fun prodColumnPowSingleThread(rows: List<VectorCiphertext>, exps: VectorQ): VectorCiphertext {
-    val nrows = rows.size
-    require(exps.nelems == nrows)
-    val width = rows[0].nelems
-    val result = List(width) { col ->
-        val column = List(nrows) { row -> rows[row].elems[col] }
-        val columnV = VectorCiphertext(exps.group, column)
-        Prod(columnV powP exps) // CE 2 * n * width exp
-    }
-    return VectorCiphertext(exps.group, result)
-}
-
-fun calcOneCol(columnV: VectorCiphertext, exps: VectorQ): ElGamalCiphertext {
-    require(exps.nelems == columnV.nelems)
-    return Prod(columnV powP exps) // CE 2 * width exp
-}
-
-// parellel calculator of product of columns vectors to a power
-class PprodColumnPow(val rows: List<VectorCiphertext>, val exps: VectorQ, val nthreads: Int = 10) {
-    val group: GroupContext = exps.group
-    val results = mutableMapOf<Int, ElGamalCiphertext>()
-
-    fun calc(): VectorCiphertext {
-        require(exps.nelems == rows.size)
-
-        runBlocking {
-            val jobs = mutableListOf<Job>()
-            val colProducer = producer(rows)
-            repeat(nthreads) {
-                jobs.add(launchCalculator(colProducer) { (columnV, colIdx) ->
-                    Pair(calcOneCol(columnV, exps), colIdx)
-                })
-            }
-            // wait for all calculations to be done, then close everything
-            joinAll(*jobs.toTypedArray())
-        }
-
-        // put results in order
-        val columns = List(results.size) { results[it]!! }
-        return VectorCiphertext(group, columns)
-    }
-
-    private fun CoroutineScope.producer(rows: List<VectorCiphertext>): ReceiveChannel< Pair<VectorCiphertext, Int> > =
-        produce {
-            val nrows = rows.size
-            val width = rows[0].nelems
-            List(width) { col ->
-                val column = List(nrows) { row -> rows[row].elems[col] }
-                val columnV = VectorCiphertext(exps.group, column)
-                send(Pair(columnV, col))
-                yield()
-            }
-            channel.close()
-        }
-
-    private val mutex = Mutex()
-
-    private fun CoroutineScope.launchCalculator(
-        producer: ReceiveChannel<Pair<VectorCiphertext, Int>>,
-        calculate: (Pair<VectorCiphertext, Int>) -> Pair<ElGamalCiphertext, Int>
-    ) = launch(Dispatchers.Default) {
-
-        for (pair in producer) {
-            val (column, idx) = calculate(pair)
-            mutex.withLock {
-                results[idx] = column
-            }
-            yield()
-        }
-    }
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 
 // parallel computation of B and Bp
