+++ src/test/java/org/cryptobiotic/bigint/DoubleConsts.java
