+++ src/test/java/org/cryptobiotic/bigint/FloatConsts.java
